refactor: share formatting helpers across comparison reports
